feat: Use forge install instead of specifying forge-std repo
This commit changes the forge install command in the CI workflow to use `forge install` instead of `forge install foundry-rs/forge-std --no-commit`. This is because forge-std is now included as a submodule, so it can be installed with the default command. This simplifies the CI workflow and ensures that the correct version of forge-std is used.