Refactor: Rename contracts to V1 and add V2
This commit renames `MutatedOptionFactory.sol` and `MutatedOptionPair.sol` to `MutatedOptionFactoryV1.sol` and `MutatedOptionPairV1.sol` respectively.
It also introduces a new `MutatedOptionPairV2.sol` contract.
This change is part of a larger effort to introduce a new version of the MutatedOptionPair contract.
The tests have been updated to reflect the new contract names.